handle random devices on fread
When intercepting fread, we now check if the fd corresponds to a
random device, and if so, Shadow returns random bytes from its own
random source rather than allowing the kernel to return random
bytes. This is done to ensure deterministic experiments.

refs #368

+++ src/main/utility/shd-random.c
@@ -48,12 +48,12 @@ guint random_nextUInt(Random* random) {
     return (guint)randomUint;
 }
 
-void random_nextNBytes(Random* random, guchar* buffer, gint nbytes) {
+void random_nextNBytes(Random* random, guchar* buffer, gsize nbytes) {
     utility_assert(random);
-    gint offset = 0;
+    gsize offset = 0;
     while(offset < nbytes) {
         guint randUInt = random_nextUInt(random);
-        gint n = MIN((nbytes - offset), sizeof(guint));
+        gsize n = MIN((nbytes - offset), sizeof(guint));
         memmove(&buffer[offset], &randUInt, n);
         offset += n;
     }
