+++ src/main/utility/shd-random.h
@@ -47,6 +47,6 @@ guint random_nextUInt(Random* random);
  * @param buffer the buffer to copy the random bytes to
  * @param nbytes number of bytes to copy to the buffer
  */
-void random_nextNBytes(Random* random, guchar* buffer, gint nbytes);
+void random_nextNBytes(Random* random, guchar* buffer, gsize nbytes);
 
 #endif /* SHD_RANDOM_H_ */
